Revamped gee download structure and code
