+++ ui/gee.ui
@@ -215,14 +215,27 @@
       <widget class="QPushButton" name="downloadBtn">
        <property name="geometry">
         <rect>
-         <x>220</x>
-         <y>400</y>
+         <x>120</x>
+         <y>390</y>
          <width>113</width>
          <height>32</height>
         </rect>
        </property>
        <property name="text">
         <string>Download</string>
+       </property>
+      </widget>
+      <widget class="QProgressBar" name="downloadProgressBar">
+       <property name="geometry">
+        <rect>
+         <x>250</x>
+         <y>390</y>
+         <width>118</width>
+         <height>31</height>
+        </rect>
+       </property>
+       <property name="value">
+        <number>0</number>
        </property>
       </widget>
      </widget>
